backend/pyproject.toml
[tool.poetry]
name = "survey-app-backend"
version = "0.1.0"
description = "Backend part for survey app course work"
authors = ["DragonProd"]


[tool.poetry.dependencies]
python = "^3.9.0"
alembic = "^1.8.1"
fastapi = "^0.79.0"
SQLAlchemy = "^1.4.40"
uvicorn = "^0.18.2"
python-dotenv = "^0.20.0"
asyncpg = "^0.26.0"
python-multipart = "^0.0.5"
Pillow = "^9.2.0"
httpx = "^0.23.0"
python-jose = "^3.3.0"
email-validator = "^1.1.3"
pydantic = {extras = ["email"], version = "^1.9.2"}
aiohttp = "^3.8.1"
passlib = "^1.7.4"
loguru = "^0.6.0"
requests = "^2.28.1"


[tool.poetry.dev-dependencies]
mypy = "^0.942"
black = {extras = ["d"], version = "^22.8.0"}
isort = "^5.10.1"
autoflake = "^1.4"
flake8 = "^4.0.1"
pytest = "^7.1.1"
sqlalchemy-stubs = "^0.4"
pytest-cov = "^3.0.0"
pytest-alembic = "^0.8.4"
pytest-asyncio = "^0.19.0"
pytest-trio = "^0.7.0"


[tool.black]
line-length = 120
target-version = ['py39']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
)
'''

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 120


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
